fix: resolve CI failure and modularize system components
- Disable CodeQL analysis in security workflow as it is not enabled for the repository.
- Modularize worker/lib/logger.ts into worker/lib/logger/ directory to meet the 500-line file size limit.
- Fix XSS vulnerabilities in extension/popup.js by replacing innerHTML with textContent.
- Refactor worker/index.ts and worker/routes/core.ts to improve modularity and satisfy TypeScript checks.
- Address Node.js deprecation warnings in GitHub Action workflows by opting into Node 24 actions.
- Ensure all unit and integration tests pass locally.

Addresses Issue #21 and production readiness items.
