+++ .github/workflows/deploy-production.yml
@@ -1,5 +1,8 @@
 name: Deploy - Production
 
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
+
 on:
   push:
     branches: [main]
